Structural refactoring
- This is in preparation for multi version support.
- The API has to be properly reworked to expose unnecessarily deep
module structs (re-export useful structs).
