+++ src/action/chunk.rs
@@ -1,0 +1,73 @@
+#[cfg(test)]
+use crate::test_tools::test_item;
+#[cfg(test)]
+use hex_literal::hex;
+
+use crate::{
+    action::OpCode,
+    codec::{Codec, StdError, WithOffset, WithSize},
+};
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum ChunkStep {
+    Continue = 0,
+    Start = 1,
+    End = 2,
+    StartEnd = 3,
+}
+impl ChunkStep {
+    fn from(n: u8) -> Self {
+        match n {
+            0 => ChunkStep::Continue,
+            1 => ChunkStep::Start,
+            2 => ChunkStep::End,
+            3 => ChunkStep::StartEnd,
+            x => panic!("Impossible chunk step {}", x),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct Chunk {
+    pub step: ChunkStep,
+}
+impl std::fmt::Display for Chunk {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match &self.step {
+            ChunkStep::Continue => write!(f, "[C]"),
+            ChunkStep::Start => write!(f, "[S]"),
+            ChunkStep::End => write!(f, "[E]"),
+            ChunkStep::StartEnd => write!(f, "[R]"),
+        }
+    }
+}
+impl Codec for Chunk {
+    type Error = StdError;
+    fn encoded_size(&self) -> usize {
+        1
+    }
+    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
+        out[0] = OpCode::Chunk as u8 + ((self.step as u8) << 6);
+        1
+    }
+    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
+        if out.is_empty() {
+            return Err(WithOffset::new_head(Self::Error::MissingBytes(1)));
+        }
+        Ok(WithSize {
+            value: Self {
+                step: ChunkStep::from(out[0] >> 6),
+            },
+            size: 1,
+        })
+    }
+}
+#[test]
+fn test_chunk() {
+    test_item(
+        Chunk {
+            step: ChunkStep::End,
+        },
+        &hex!("B0"),
+    )
+}
