+++ src/action/logic.rs
@@ -1,0 +1,69 @@
+#[cfg(test)]
+use crate::test_tools::test_item;
+#[cfg(test)]
+use hex_literal::hex;
+
+use crate::codec::{Codec, StdError, WithOffset, WithSize};
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum LogicOp {
+    Or = 0,
+    Xor = 1,
+    Nor = 2,
+    Nand = 3,
+}
+impl LogicOp {
+    fn from(n: u8) -> Self {
+        match n {
+            0 => LogicOp::Or,
+            1 => LogicOp::Xor,
+            2 => LogicOp::Nor,
+            3 => LogicOp::Nand,
+            x => panic!("Impossible logic op {}", x),
+        }
+    }
+}
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct Logic {
+    pub logic: LogicOp,
+}
+impl std::fmt::Display for Logic {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match &self.logic {
+            LogicOp::Or => write!(f, "[OR]"),
+            LogicOp::Xor => write!(f, "[XOR]"),
+            LogicOp::Nor => write!(f, "[NOR]"),
+            LogicOp::Nand => write!(f, "[NAND]"),
+        }
+    }
+}
+impl Codec for Logic {
+    type Error = StdError;
+    fn encoded_size(&self) -> usize {
+        1
+    }
+    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
+        out[0] = crate::action::OpCode::Logic as u8 + ((self.logic as u8) << 6);
+        1
+    }
+    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
+        if out.is_empty() {
+            return Err(WithOffset::new_head(Self::Error::MissingBytes(1)));
+        }
+        Ok(WithSize {
+            value: Self {
+                logic: LogicOp::from(out[0] >> 6),
+            },
+            size: 1,
+        })
+    }
+}
+#[test]
+fn test_logic() {
+    test_item(
+        Logic {
+            logic: LogicOp::Nand,
+        },
+        &hex!("F1"),
+    )
+}
