+++ src/action/response_tag.rs
@@ -1,0 +1,42 @@
+#[cfg(test)]
+use crate::test_tools::test_item;
+#[cfg(test)]
+use hex_literal::hex;
+
+use crate::codec::{Codec, StdError, WithOffset, WithSize};
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct ResponseTag {
+    /// End of packet
+    ///
+    /// Signal the last response packet for the request `id`
+    /// (E)
+    pub eop: bool,
+    /// An error occured
+    /// (R)
+    pub err: bool,
+    pub id: u8,
+}
+impl std::fmt::Display for ResponseTag {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(
+            f,
+            "[{}{}]({})",
+            if self.eop { "E" } else { "-" },
+            if self.err { "R" } else { "-" },
+            self.id,
+        )
+    }
+}
+super::impl_simple_op!(ResponseTag, eop, err, id);
+#[test]
+fn test_response_tag() {
+    test_item(
+        ResponseTag {
+            eop: true,
+            err: false,
+            id: 8,
+        },
+        &hex!("A3 08"),
+    )
+}
