+++ src/action/write_file_properties.rs
@@ -1,0 +1,68 @@
+#[cfg(test)]
+use crate::test_tools::test_item;
+#[cfg(test)]
+use hex_literal::hex;
+
+use crate::{
+    codec::{Codec, WithOffset, WithSize},
+    data,
+};
+/// Write the properties of a file
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct WriteFileProperties {
+    /// Group with next action
+    pub group: bool,
+    /// Ask for a response (a status)
+    pub resp: bool,
+    pub file_id: u8,
+    pub header: data::FileHeader,
+}
+super::impl_header_op!(WriteFileProperties, group, resp, file_id, header);
+impl std::fmt::Display for WriteFileProperties {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(
+            f,
+            "[{}{}]f({}){}",
+            if self.group { "G" } else { "-" },
+            if self.resp { "R" } else { "-" },
+            self.file_id,
+            self.header,
+        )
+    }
+}
+#[test]
+fn test_write_file_properties() {
+    test_item(
+        WriteFileProperties {
+            group: true,
+            resp: false,
+            file_id: 9,
+            header: data::FileHeader {
+                permissions: data::Permissions {
+                    encrypted: true,
+                    executable: false,
+                    user: data::UserPermissions {
+                        read: true,
+                        write: true,
+                        run: true,
+                    },
+                    guest: data::UserPermissions {
+                        read: false,
+                        write: false,
+                        run: false,
+                    },
+                },
+                properties: data::FileProperties {
+                    act_en: false,
+                    act_cond: data::ActionCondition::Read,
+                    storage_class: data::StorageClass::Permanent,
+                },
+                alp_cmd_fid: 1,
+                interface_file_id: 2,
+                file_size: 0xDEAD_BEEF,
+                allocated_size: 0xBAAD_FACE,
+            },
+        },
+        &hex!("86   09   B8 13 01 02 DEADBEEF BAADFACE"),
+    )
+}
