+++ src/operand/query.rs
@@ -1,480 +1,12 @@
+use crate::operand::file_offset::{FileOffset, FileOffsetDecodingError};
 #[cfg(test)]
 use crate::test_tools::test_item;
 use crate::{
     codec::{Codec, StdError, WithOffset, WithSize},
-    dash7, varint,
+    varint,
 };
 #[cfg(test)]
 use hex_literal::hex;
-
-// ===============================================================================
-// Alp Interfaces
-// ===============================================================================
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum InterfaceId {
-    Host = 0,
-    D7asp = 0xD7,
-}
-
-/// Meta data required to send a packet depending on the sending interface type
-#[derive(Clone, Debug, PartialEq)]
-pub enum InterfaceConfiguration {
-    Host,
-    D7asp(dash7::InterfaceConfiguration),
-}
-impl std::fmt::Display for InterfaceConfiguration {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Self::Host => write!(f, "HOST"),
-            Self::D7asp(conf) => write!(f, "D7:{}", conf),
-        }
-    }
-}
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum InterfaceConfigurationDecodingError {
-    MissingBytes(usize),
-    D7asp(dash7::InterfaceConfigurationDecodingError),
-    BadInterfaceId(u8),
-}
-impl Codec for InterfaceConfiguration {
-    type Error = InterfaceConfigurationDecodingError;
-    fn encoded_size(&self) -> usize {
-        1 + match self {
-            InterfaceConfiguration::Host => 0,
-            InterfaceConfiguration::D7asp(v) => v.encoded_size(),
-        }
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        match self {
-            InterfaceConfiguration::Host => {
-                out[0] = InterfaceId::Host as u8;
-                1
-            }
-            InterfaceConfiguration::D7asp(v) => {
-                out[0] = InterfaceId::D7asp as u8;
-                1 + v.encode_in(&mut out[1..])
-            }
-        }
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.is_empty() {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(1)));
-        }
-        const HOST: u8 = InterfaceId::Host as u8;
-        const D7ASP: u8 = InterfaceId::D7asp as u8;
-        Ok(match out[0] {
-            HOST => WithSize {
-                value: InterfaceConfiguration::Host,
-                size: 1,
-            },
-            D7ASP => {
-                let WithSize { value, size } = dash7::InterfaceConfiguration::decode(&out[1..])
-                    .map_err(|e| e.map_value(InterfaceConfigurationDecodingError::D7asp))?;
-                WithSize {
-                    value: InterfaceConfiguration::D7asp(value),
-                    size: size + 1,
-                }
-            }
-            id => {
-                return Err(WithOffset {
-                    value: Self::Error::BadInterfaceId(id),
-                    offset: 0,
-                })
-            }
-        })
-    }
-}
-#[test]
-fn test_interface_configuration_d7asp() {
-    test_item(
-        InterfaceConfiguration::D7asp(dash7::InterfaceConfiguration {
-            qos: dash7::Qos {
-                retry: dash7::RetryMode::No,
-                resp: dash7::RespMode::Any,
-            },
-            to: 0x23,
-            te: 0x34,
-            nls_method: dash7::NlsMethod::AesCcm32,
-            access_class: 0xFF,
-            address: dash7::Address::Vid([0xAB, 0xCD]),
-        }),
-        &hex!("D7   02 23 34   37 FF ABCD"),
-    )
-}
-#[test]
-fn test_interface_configuration_host() {
-    test_item(InterfaceConfiguration::Host, &hex!("00"))
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct InterfaceStatusUnknown {
-    pub id: u8,
-    pub data: Box<[u8]>,
-}
-impl std::fmt::Display for InterfaceStatusUnknown {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}:0x{}", self.id, hex::encode_upper(&self.data))
-    }
-}
-/// Meta data from a received packet depending on the receiving interface type
-#[derive(Clone, Debug, PartialEq)]
-pub enum InterfaceStatus {
-    Host,
-    D7asp(dash7::InterfaceStatus),
-    Unknown(InterfaceStatusUnknown),
-}
-impl std::fmt::Display for InterfaceStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Self::Host => write!(f, "HOST"),
-            Self::D7asp(status) => write!(f, "D7={}", status),
-            Self::Unknown(status) => write!(f, "?={}", status),
-        }
-    }
-}
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum InterfaceStatusDecodingError {
-    MissingBytes(usize),
-    BadInterfaceId(u8),
-}
-impl From<StdError> for InterfaceStatusDecodingError {
-    fn from(e: StdError) -> Self {
-        match e {
-            StdError::MissingBytes(n) => Self::MissingBytes(n),
-        }
-    }
-}
-impl Codec for InterfaceStatus {
-    type Error = InterfaceStatusDecodingError;
-    fn encoded_size(&self) -> usize {
-        let data_size = match self {
-            InterfaceStatus::Host => 0,
-            InterfaceStatus::D7asp(itf) => itf.encoded_size(),
-            InterfaceStatus::Unknown(InterfaceStatusUnknown { data, .. }) => data.len(),
-        };
-        1 + unsafe { varint::size(data_size as u32) } as usize + data_size
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        let mut offset = 1;
-        match self {
-            InterfaceStatus::Host => {
-                out[0] = InterfaceId::Host as u8;
-                out[1] = 0;
-                offset += 1;
-            }
-            InterfaceStatus::D7asp(v) => {
-                out[0] = InterfaceId::D7asp as u8;
-                let size = v.encoded_size() as u32;
-                let size_size = varint::encode_in(size, &mut out[offset..]);
-                offset += size_size as usize;
-                offset += v.encode_in(&mut out[offset..]);
-            }
-            InterfaceStatus::Unknown(InterfaceStatusUnknown { id, data, .. }) => {
-                out[0] = *id;
-                let size = data.len() as u32;
-                let size_size = varint::encode_in(size, &mut out[offset..]);
-                offset += size_size as usize;
-                out[offset..offset + data.len()].clone_from_slice(data);
-                offset += data.len();
-            }
-        };
-        offset
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.is_empty() {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(1)));
-        }
-        const HOST: u8 = InterfaceId::Host as u8;
-        const D7ASP: u8 = InterfaceId::D7asp as u8;
-        let mut offset = 1;
-        let value = match out[0] {
-            HOST => {
-                offset += 1;
-                InterfaceStatus::Host
-            }
-            D7ASP => {
-                let WithSize {
-                    value: size,
-                    size: size_size,
-                } = varint::decode(&out[offset..]).map_err(|e| {
-                    let WithOffset { offset: off, value } = e;
-                    WithOffset {
-                        offset: offset + off,
-                        value: value.into(),
-                    }
-                })?;
-                let size = size as usize;
-                offset += size_size;
-                let WithSize { value, size } =
-                    dash7::InterfaceStatus::decode(&out[offset..offset + size]).map_err(|e| {
-                        let WithOffset { offset: off, value } = e;
-                        WithOffset {
-                            offset: offset + off,
-                            value: value.into(),
-                        }
-                    })?;
-                offset += size;
-                InterfaceStatus::D7asp(value)
-            }
-            id => {
-                let WithSize {
-                    value: size,
-                    size: size_size,
-                } = varint::decode(&out[offset..]).map_err(|e| {
-                    let WithOffset { offset: off, value } = e;
-                    WithOffset {
-                        offset: offset + off,
-                        value: value.into(),
-                    }
-                })?;
-                let size = size as usize;
-                offset += size_size;
-                if out.len() < offset + size {
-                    return Err(WithOffset::new(
-                        offset,
-                        Self::Error::MissingBytes(offset + size - out.len()),
-                    ));
-                }
-                let mut data = vec![0u8; size].into_boxed_slice();
-                data.clone_from_slice(&out[offset..size]);
-                offset += size;
-                InterfaceStatus::Unknown(InterfaceStatusUnknown { id, data })
-            }
-        };
-        Ok(WithSize {
-            value,
-            size: offset,
-        })
-    }
-}
-#[test]
-fn test_interface_status_d7asp() {
-    test_item(
-        InterfaceStatus::D7asp(dash7::InterfaceStatus {
-            ch_header: 1,
-            ch_idx: 0x0123,
-            rxlev: 2,
-            lb: 3,
-            snr: 4,
-            status: 0xB0,
-            token: 6,
-            seq: 7,
-            resp_to: 8,
-            access_class: 0xFF,
-            address: dash7::Address::Vid([0xAB, 0xCD]),
-            nls_state: dash7::NlsState::AesCcm32(hex!("00 11 22 33 44")),
-        }),
-        &hex!("D7 13    01 0123 02 03 04 B0 06 07 08   37 FF ABCD  0011223344"),
-    )
-}
-#[test]
-fn test_interface_status_host() {
-    test_item(InterfaceStatus::Host, &hex!("00 00"))
-}
-
-// ===============================================================================
-// Operands
-// ===============================================================================
-/// Describe the location of some data on the filesystem (file + data offset).
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct FileOffset {
-    pub id: u8,
-    pub offset: u32,
-}
-impl std::fmt::Display for FileOffset {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{},{}", self.id, self.offset)
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum FileOffsetDecodingError {
-    MissingBytes(usize),
-    Offset(StdError),
-}
-impl Codec for FileOffset {
-    type Error = FileOffsetDecodingError;
-    fn encoded_size(&self) -> usize {
-        1 + unsafe { varint::size(self.offset) } as usize
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        out[0] = self.id;
-        1 + varint::encode_in(self.offset, &mut out[1..]) as usize
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.len() < 2 {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(
-                2 - out.len(),
-            )));
-        }
-        let WithSize {
-            value: offset,
-            size,
-        } = varint::decode(&out[1..]).map_err(|e| {
-            let WithOffset { offset, value } = e;
-            WithOffset {
-                offset: offset + 1,
-                value: FileOffsetDecodingError::Offset(value),
-            }
-        })?;
-        Ok(WithSize {
-            value: Self { id: out[0], offset },
-            size: 1 + size,
-        })
-    }
-}
-#[test]
-fn test_file_offset_operand() {
-    test_item(
-        FileOffset {
-            id: 2,
-            offset: 0x3F_FF,
-        },
-        &hex!("02 7F FF"),
-    )
-}
-
-pub mod status_code {
-    //! Status code that can be received as a result of some ALP actions.
-    /// Action received and partially completed at response. To be completed after response
-    pub const RECEIVED: u8 = 1;
-    pub const OK: u8 = 0;
-    pub const FILE_ID_MISSING: u8 = 0xFF;
-    pub const CREATE_FILE_ID_ALREADY_EXIST: u8 = 0xFE;
-    pub const FILE_IS_NOT_RESTORABLE: u8 = 0xFD;
-    pub const INSUFFICIENT_PERMISSION: u8 = 0xFC;
-    pub const CREATE_FILE_LENGTH_OVERFLOW: u8 = 0xFB;
-    pub const CREATE_FILE_ALLOCATION_OVERFLOW: u8 = 0xFA; // ALP_SPEC: ??? Difference with the previous one?;
-    pub const WRITE_OFFSET_OVERFLOW: u8 = 0xF9;
-    pub const WRITE_DATA_OVERFLOW: u8 = 0xF8;
-    pub const WRITE_STORAGE_UNAVAILABLE: u8 = 0xF7;
-    pub const UNKNOWN_OPERATION: u8 = 0xF6;
-    pub const OPERAND_INCOMPLETE: u8 = 0xF5;
-    pub const OPERAND_WRONG_FORMAT: u8 = 0xF4;
-    pub const UNKNOWN_ERROR: u8 = 0x80;
-}
-
-/// Result of an action in a previously sent request
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct Status {
-    /// Index of the ALP action associated with this status, in the original request as seen from
-    /// the receiver side.
-    // ALP_SPEC This is complicated to process because we have to known/possibly infer the position
-    // of the action on the receiver side, and that we have to do that while also interpreting who
-    // responded (the local modem won't have the same index as the distant device.).
-    pub action_id: u8,
-    /// Result code
-    pub status: u8,
-}
-impl std::fmt::Display for Status {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "a[{}]=>{}", self.action_id, self.status)
-    }
-}
-impl Codec for Status {
-    type Error = StdError;
-    fn encoded_size(&self) -> usize {
-        1 + 1
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        out[0] = self.action_id;
-        out[1] = self.status as u8;
-        2
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.len() < 2 {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(
-                2 - out.len(),
-            )));
-        }
-        Ok(WithSize {
-            value: Self {
-                action_id: out[0],
-                status: out[1],
-            },
-            size: 2,
-        })
-    }
-}
-#[test]
-fn test_status_operand() {
-    test_item(
-        Status {
-            action_id: 2,
-            status: status_code::UNKNOWN_OPERATION,
-        },
-        &hex!("02 F6"),
-    )
-}
-
-// ALP SPEC: where is this defined? Link? Not found in either specs !
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum Permission {
-    Dash7([u8; 8]),
-}
-
-impl Permission {
-    fn id(self) -> u8 {
-        match self {
-            Permission::Dash7(_) => 0x42, // ALP_SPEC Undefined
-        }
-    }
-}
-impl std::fmt::Display for Permission {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Self::Dash7(data) => write!(f, "D7:0x{}", hex::encode_upper(data)),
-        }
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum PermissionDecodingError {
-    MissingBytes(usize),
-    UnknownId(u8),
-}
-
-impl Codec for Permission {
-    type Error = PermissionDecodingError;
-    fn encoded_size(&self) -> usize {
-        1 + match self {
-            Permission::Dash7(_) => 8,
-        }
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        out[0] = self.id();
-        1 + match self {
-            Permission::Dash7(token) => {
-                out[1..1 + token.len()].clone_from_slice(&token[..]);
-                8
-            }
-        }
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.is_empty() {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(1)));
-        }
-        let mut offset = 1;
-        match out[0] {
-            0x42 => {
-                let mut token = [0; 8];
-                token.clone_from_slice(&out[offset..offset + 8]);
-                offset += 8;
-                Ok(WithSize {
-                    value: Permission::Dash7(token),
-                    size: offset,
-                })
-            }
-            x => Err(WithOffset::new_head(Self::Error::UnknownId(x))),
-        }
-    }
-}
-
-pub mod permission_level {
-    pub const USER: u8 = 0;
-    pub const ROOT: u8 = 1;
-    // ALP SPEC: Does something else exist?
-}
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum QueryComparisonType {
@@ -1448,161 +980,5 @@ impl Codec for Query {
                 .map(|ok| ok.map_value(Query::StringTokenSearch))
                 .map_err(|e| e.map_value(Self::Error::StringTokenSearch)),
         }
-    }
-}
-
-/// Dash7 interface
-#[derive(Clone, Debug, PartialEq)]
-pub struct OverloadedIndirectInterface {
-    /// File containing the `QoS`, `to` and `te` to use for the transmission (see
-    /// dash7::InterfaceConfiguration
-    pub interface_file_id: u8,
-    pub nls_method: dash7::NlsMethod,
-    pub access_class: u8,
-    pub address: dash7::Address,
-}
-impl std::fmt::Display for OverloadedIndirectInterface {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{},{},{},{}",
-            self.interface_file_id, self.nls_method, self.access_class, self.address
-        )
-    }
-}
-
-impl Codec for OverloadedIndirectInterface {
-    type Error = StdError;
-    fn encoded_size(&self) -> usize {
-        1 + 2 + self.address.encoded_size()
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        out[0] = self.interface_file_id;
-        out[1] = ((self.address.id_type() as u8) << 4) | (self.nls_method as u8);
-        out[2] = self.access_class;
-        1 + 2 + self.address.encode_in(&mut out[3..])
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.len() < 1 + 2 {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(
-                1 + 2 - out.len(),
-            )));
-        }
-        let interface_file_id = out[0];
-        let address_type = dash7::AddressType::from((out[1] & 0x30) >> 4);
-        let nls_method = unsafe { dash7::NlsMethod::from(out[1] & 0x0F) };
-        let access_class = out[2];
-        let WithSize {
-            value: address,
-            size: address_size,
-        } = dash7::Address::parse(address_type, &out[3..]).map_err(|e| e.shift(3))?;
-        Ok(WithSize {
-            value: Self {
-                interface_file_id,
-                nls_method,
-                access_class,
-                address,
-            },
-            size: 1 + 2 + address_size,
-        })
-    }
-}
-#[test]
-fn test_overloaded_indirect_interface() {
-    test_item(
-        OverloadedIndirectInterface {
-            interface_file_id: 4,
-            nls_method: dash7::NlsMethod::AesCcm32,
-            access_class: 0xFF,
-            address: dash7::Address::Vid([0xAB, 0xCD]),
-        },
-        &hex!("04   37 FF ABCD"),
-    )
-}
-
-/// Non Dash7 interface
-#[derive(Clone, Debug, PartialEq)]
-// ALP SPEC: This seems undoable if we do not know the interface (per protocol specific support)
-//  which is still a pretty legitimate policy on a low power protocol.
-pub struct NonOverloadedIndirectInterface {
-    pub interface_file_id: u8,
-    // ALP SPEC: Where is this defined? Is this ID specific?
-    pub data: Box<[u8]>,
-}
-
-impl Codec for NonOverloadedIndirectInterface {
-    type Error = StdError;
-    fn encoded_size(&self) -> usize {
-        1 + self.data.len()
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        out[0] = self.interface_file_id;
-        let mut offset = 1;
-        out[offset..offset + self.data.len()].clone_from_slice(&self.data);
-        offset += self.data.len();
-        offset
-    }
-    fn decode(_out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        todo!("TODO")
-    }
-}
-impl std::fmt::Display for NonOverloadedIndirectInterface {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{},0x{}",
-            self.interface_file_id,
-            hex::encode_upper(&self.data)
-        )
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum IndirectInterface {
-    Overloaded(OverloadedIndirectInterface),
-    NonOverloaded(NonOverloadedIndirectInterface),
-}
-impl std::fmt::Display for IndirectInterface {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Self::Overloaded(v) => write!(f, "O:{}", v),
-            Self::NonOverloaded(v) => write!(f, "N:{}", v),
-        }
-    }
-}
-
-impl Codec for IndirectInterface {
-    type Error = StdError;
-    fn encoded_size(&self) -> usize {
-        match self {
-            IndirectInterface::Overloaded(v) => v.encoded_size(),
-            IndirectInterface::NonOverloaded(v) => v.encoded_size(),
-        }
-    }
-    unsafe fn encode_in(&self, out: &mut [u8]) -> usize {
-        match self {
-            IndirectInterface::Overloaded(v) => v.encode_in(out),
-            IndirectInterface::NonOverloaded(v) => v.encode_in(out),
-        }
-    }
-    fn decode(out: &[u8]) -> Result<WithSize<Self>, WithOffset<Self::Error>> {
-        if out.is_empty() {
-            return Err(WithOffset::new_head(Self::Error::MissingBytes(1)));
-        }
-        Ok(if out[0] & 0x80 != 0 {
-            let WithSize { size, value } =
-                OverloadedIndirectInterface::decode(&out[1..]).map_err(|e| e.shift(1))?;
-            WithSize {
-                size: size + 1,
-                value: Self::Overloaded(value),
-            }
-        } else {
-            let WithSize { size, value } =
-                NonOverloadedIndirectInterface::decode(&out[1..]).map_err(|e| e.shift(1))?;
-            WithSize {
-                size: size + 1,
-                value: Self::NonOverloaded(value),
-            }
-        })
     }
 }
